+++ PersistenciaPI4/src/br/senac/model/EAO/impl/AcessorioEAOImpl.java
@@ -54,13 +54,13 @@ public class AcessorioEAOImpl implements AcessorioEAO {
     }
 
     @Override
-    public Acessorio getAcessorio(Integer id) {
+    public Acessorio getAcessorio(Long id) {
         entityManager = dbSingleton.getEntityManager();               
         return (Acessorio) entityManager.find(Acessorio.class, id);
     }
 
     @Override
-    public boolean deletar(Integer id) {
+    public boolean deletar(Long id) {
         entityManager = dbSingleton.getEntityManager();
         try{            
             entityManager.getTransaction().begin();
