+++ PersistenciaPI4/src/br/senac/test/AcessorioTest.java
@@ -28,7 +28,7 @@ public class AcessorioTest {
 	}
 	
 	
-	public static void testaBuscarPorId(int id){
+	public static void testaBuscarPorId(Long id){
 		AcessorioEAOImpl acessorioEAO = new AcessorioEAOImpl();
 		Acessorio acessorio = new Acessorio();
 		acessorio = acessorioEAO.getAcessorio(id);
@@ -39,17 +39,17 @@ public class AcessorioTest {
 	public static void testaEditar(){
 		Acessorio acessorio = new Acessorio();		
 		AcessorioEAOImpl acessorioEAO = new AcessorioEAOImpl();
-		acessorio = acessorioEAO.getAcessorio(5);
+		acessorio = acessorioEAO.getAcessorio(5L);
 		System.out.println("Antigo: "+acessorio.getDescricao());
 		acessorio.setDescricao("Descrição alterada!");
 		acessorioEAO.editar(acessorio);
-		acessorio = acessorioEAO.getAcessorio(5);
+		acessorio = acessorioEAO.getAcessorio(5L);
 		System.out.println("Nova: "+acessorio.getDescricao());
 	}
         
         public static void testaRemover(){
             AcessorioEAOImpl acessorioEAOImpl = new AcessorioEAOImpl();
-            acessorioEAOImpl.deletar(6);
+            acessorioEAOImpl.deletar(6L);
         }
 	
 }
