+++ pi4-app/src/main/java/br/senac/service/AcessorioService.java
@@ -12,8 +12,8 @@ public interface AcessorioService {
 
 	public void editar(Acessorio acessorio);
 
-	public Acessorio getAcessorio(Integer id);
+	public Acessorio getAcessorio(Long id);
 
-	public boolean deletar(Integer id);
+	public boolean deletar(Long id);
 
 }
